+++ packages/components-angular/src/app/pages/index.ts
@@ -70,11 +70,23 @@ import { DividerComponent } from './generated/divider.component';
 import { FieldsetComponent } from './generated/fieldset.component';
 import { FieldsetWrapperComponent } from './generated/fieldset-wrapper.component';
 import { FlexComponent } from './generated/flex.component';
-import { FlyoutComponent } from './generated/flyout.component';
+import { Flyout_1Component } from './generated/flyout-1.component';
+import { Flyout_2Component } from './generated/flyout-2.component';
 import { FlyoutBasicComponent } from './generated/flyout-basic.component';
 import { FlyoutCssVariablesComponent } from './generated/flyout-css-variables.component';
+import { FlyoutFixedFooterComponent } from './generated/flyout-fixed-footer.component';
+import { FlyoutFixedFooterHeaderComponent } from './generated/flyout-fixed-footer-header.component';
+import { FlyoutFixedFooterHeaderScrollableContentComponent } from './generated/flyout-fixed-footer-header-scrollable-content.component';
+import { FlyoutFixedFooterHeaderSubfooterComponent } from './generated/flyout-fixed-footer-header-subfooter.component';
+import { FlyoutFixedFooterHeaderSubfooterScrollableContentComponent } from './generated/flyout-fixed-footer-header-subfooter-scrollable-content.component';
+import { FlyoutFixedFooterScrollableContentComponent } from './generated/flyout-fixed-footer-scrollable-content.component';
+import { FlyoutFixedFooterScrollableContentScrolledComponent } from './generated/flyout-fixed-footer-scrollable-content-scrolled.component';
+import { FlyoutFixedFooterSubfooterComponent } from './generated/flyout-fixed-footer-subfooter.component';
+import { FlyoutFixedFooterSubfooterScrollableContentComponent } from './generated/flyout-fixed-footer-subfooter-scrollable-content.component';
+import { FlyoutFixedFooterSubfooterScrollableContentScrolledComponent } from './generated/flyout-fixed-footer-subfooter-scrollable-content-scrolled.component';
 import { FlyoutFooterComponent } from './generated/flyout-footer.component';
 import { FlyoutFooterScrollableContentComponent } from './generated/flyout-footer-scrollable-content.component';
+import { FlyoutFooterScrolledComponent } from './generated/flyout-footer-scrolled.component';
 import { FlyoutGridComponent } from './generated/flyout-grid.component';
 import { FlyoutHeaderComponent } from './generated/flyout-header.component';
 import { FlyoutHeaderFooterComponent } from './generated/flyout-header-footer.component';
@@ -114,6 +126,7 @@ import { ModalBackdropShadingComponent } from './generated/modal-backdrop-shadin
 import { ModalBasicComponent } from './generated/modal-basic.component';
 import { ModalCssVariablesComponent } from './generated/modal-css-variables.component';
 import { ModalDisableCloseButtonComponent } from './generated/modal-disable-close-button.component';
+import { ModalFooterScrolledComponent } from './generated/modal-footer-scrolled.component';
 import { ModalFullscreenComponent } from './generated/modal-fullscreen.component';
 import { ModalFullscreenResponsiveComponent } from './generated/modal-fullscreen-responsive.component';
 import { ModalGridComponent } from './generated/modal-grid.component';
@@ -203,11 +216,23 @@ export const generatedPages = [
   FieldsetComponent,
   FieldsetWrapperComponent,
   FlexComponent,
-  FlyoutComponent,
+  Flyout_1Component,
+  Flyout_2Component,
   FlyoutBasicComponent,
   FlyoutCssVariablesComponent,
+  FlyoutFixedFooterComponent,
+  FlyoutFixedFooterHeaderComponent,
+  FlyoutFixedFooterHeaderScrollableContentComponent,
+  FlyoutFixedFooterHeaderSubfooterComponent,
+  FlyoutFixedFooterHeaderSubfooterScrollableContentComponent,
+  FlyoutFixedFooterScrollableContentComponent,
+  FlyoutFixedFooterScrollableContentScrolledComponent,
+  FlyoutFixedFooterSubfooterComponent,
+  FlyoutFixedFooterSubfooterScrollableContentComponent,
+  FlyoutFixedFooterSubfooterScrollableContentScrolledComponent,
   FlyoutFooterComponent,
   FlyoutFooterScrollableContentComponent,
+  FlyoutFooterScrolledComponent,
   FlyoutGridComponent,
   FlyoutHeaderComponent,
   FlyoutHeaderFooterComponent,
@@ -247,6 +272,7 @@ export const generatedPages = [
   ModalBasicComponent,
   ModalCssVariablesComponent,
   ModalDisableCloseButtonComponent,
+  ModalFooterScrolledComponent,
   ModalFullscreenComponent,
   ModalFullscreenResponsiveComponent,
   ModalGridComponent,
@@ -489,9 +515,14 @@ export const generatedRoutes: ExtendedRoute[] = [
     component: FlexComponent,
   },
   {
-    name: 'Flyout',
-    path: 'flyout',
-    component: FlyoutComponent,
+    name: 'Flyout 1',
+    path: 'flyout-1',
+    component: Flyout_1Component,
+  },
+  {
+    name: 'Flyout 2',
+    path: 'flyout-2',
+    component: Flyout_2Component,
   },
   {
     name: 'Flyout Basic',
@@ -504,6 +535,56 @@ export const generatedRoutes: ExtendedRoute[] = [
     component: FlyoutCssVariablesComponent,
   },
   {
+    name: 'Flyout Fixed Footer',
+    path: 'flyout-fixed-footer',
+    component: FlyoutFixedFooterComponent,
+  },
+  {
+    name: 'Flyout Fixed Footer Header',
+    path: 'flyout-fixed-footer-header',
+    component: FlyoutFixedFooterHeaderComponent,
+  },
+  {
+    name: 'Flyout Fixed Footer Header Scrollable Content',
+    path: 'flyout-fixed-footer-header-scrollable-content',
+    component: FlyoutFixedFooterHeaderScrollableContentComponent,
+  },
+  {
+    name: 'Flyout Fixed Footer Header Subfooter',
+    path: 'flyout-fixed-footer-header-subfooter',
+    component: FlyoutFixedFooterHeaderSubfooterComponent,
+  },
+  {
+    name: 'Flyout Fixed Footer Header Subfooter Scrollable Content',
+    path: 'flyout-fixed-footer-header-subfooter-scrollable-content',
+    component: FlyoutFixedFooterHeaderSubfooterScrollableContentComponent,
+  },
+  {
+    name: 'Flyout Fixed Footer Scrollable Content',
+    path: 'flyout-fixed-footer-scrollable-content',
+    component: FlyoutFixedFooterScrollableContentComponent,
+  },
+  {
+    name: 'Flyout Fixed Footer Scrollable Content Scrolled',
+    path: 'flyout-fixed-footer-scrollable-content-scrolled',
+    component: FlyoutFixedFooterScrollableContentScrolledComponent,
+  },
+  {
+    name: 'Flyout Fixed Footer Subfooter',
+    path: 'flyout-fixed-footer-subfooter',
+    component: FlyoutFixedFooterSubfooterComponent,
+  },
+  {
+    name: 'Flyout Fixed Footer Subfooter Scrollable Content',
+    path: 'flyout-fixed-footer-subfooter-scrollable-content',
+    component: FlyoutFixedFooterSubfooterScrollableContentComponent,
+  },
+  {
+    name: 'Flyout Fixed Footer Subfooter Scrollable Content Scrolled',
+    path: 'flyout-fixed-footer-subfooter-scrollable-content-scrolled',
+    component: FlyoutFixedFooterSubfooterScrollableContentScrolledComponent,
+  },
+  {
     name: 'Flyout Footer',
     path: 'flyout-footer',
     component: FlyoutFooterComponent,
@@ -512,6 +593,11 @@ export const generatedRoutes: ExtendedRoute[] = [
     name: 'Flyout Footer Scrollable Content',
     path: 'flyout-footer-scrollable-content',
     component: FlyoutFooterScrollableContentComponent,
+  },
+  {
+    name: 'Flyout Footer Scrolled',
+    path: 'flyout-footer-scrolled',
+    component: FlyoutFooterScrolledComponent,
   },
   {
     name: 'Flyout Grid',
@@ -707,6 +793,11 @@ export const generatedRoutes: ExtendedRoute[] = [
     name: 'Modal Disable Close Button',
     path: 'modal-disable-close-button',
     component: ModalDisableCloseButtonComponent,
+  },
+  {
+    name: 'Modal Footer Scrolled',
+    path: 'modal-footer-scrolled',
+    component: ModalFooterScrolledComponent,
   },
   {
     name: 'Modal Fullscreen',