+++ packages/components-react/src/pages/index.tsx
@@ -51,11 +51,23 @@ import { DividerPage } from './generated/Divider';
 import { FieldsetPage } from './generated/Fieldset';
 import { FieldsetWrapperPage } from './generated/FieldsetWrapper';
 import { FlexPage } from './generated/Flex';
-import { FlyoutPage } from './generated/Flyout';
+import { Flyout_1Page } from './generated/Flyout_1';
+import { Flyout_2Page } from './generated/Flyout_2';
 import { FlyoutBasicPage } from './generated/FlyoutBasic';
 import { FlyoutCssVariablesPage } from './generated/FlyoutCssVariables';
+import { FlyoutFixedFooterPage } from './generated/FlyoutFixedFooter';
+import { FlyoutFixedFooterHeaderPage } from './generated/FlyoutFixedFooterHeader';
+import { FlyoutFixedFooterHeaderScrollableContentPage } from './generated/FlyoutFixedFooterHeaderScrollableContent';
+import { FlyoutFixedFooterHeaderSubfooterPage } from './generated/FlyoutFixedFooterHeaderSubfooter';
+import { FlyoutFixedFooterHeaderSubfooterScrollableContentPage } from './generated/FlyoutFixedFooterHeaderSubfooterScrollableContent';
+import { FlyoutFixedFooterScrollableContentPage } from './generated/FlyoutFixedFooterScrollableContent';
+import { FlyoutFixedFooterScrollableContentScrolledPage } from './generated/FlyoutFixedFooterScrollableContentScrolled';
+import { FlyoutFixedFooterSubfooterPage } from './generated/FlyoutFixedFooterSubfooter';
+import { FlyoutFixedFooterSubfooterScrollableContentPage } from './generated/FlyoutFixedFooterSubfooterScrollableContent';
+import { FlyoutFixedFooterSubfooterScrollableContentScrolledPage } from './generated/FlyoutFixedFooterSubfooterScrollableContentScrolled';
 import { FlyoutFooterPage } from './generated/FlyoutFooter';
 import { FlyoutFooterScrollableContentPage } from './generated/FlyoutFooterScrollableContent';
+import { FlyoutFooterScrolledPage } from './generated/FlyoutFooterScrolled';
 import { FlyoutGridPage } from './generated/FlyoutGrid';
 import { FlyoutHeaderPage } from './generated/FlyoutHeader';
 import { FlyoutHeaderFooterPage } from './generated/FlyoutHeaderFooter';
@@ -95,6 +107,7 @@ import { ModalBackdropShadingPage } from './generated/ModalBackdropShading';
 import { ModalBasicPage } from './generated/ModalBasic';
 import { ModalCssVariablesPage } from './generated/ModalCssVariables';
 import { ModalDisableCloseButtonPage } from './generated/ModalDisableCloseButton';
+import { ModalFooterScrolledPage } from './generated/ModalFooterScrolled';
 import { ModalFullscreenPage } from './generated/ModalFullscreen';
 import { ModalFullscreenResponsivePage } from './generated/ModalFullscreenResponsive';
 import { ModalGridPage } from './generated/ModalGrid';
@@ -335,9 +348,14 @@ export const generatedRoutes: RouteType[] = [
     element: <FlexPage />,
   },
   {
-    name: 'Flyout',
-    path: '/flyout',
-    element: <FlyoutPage />,
+    name: 'Flyout 1',
+    path: '/flyout-1',
+    element: <Flyout_1Page />,
+  },
+  {
+    name: 'Flyout 2',
+    path: '/flyout-2',
+    element: <Flyout_2Page />,
   },
   {
     name: 'Flyout Basic',
@@ -350,6 +368,56 @@ export const generatedRoutes: RouteType[] = [
     element: <FlyoutCssVariablesPage />,
   },
   {
+    name: 'Flyout Fixed Footer',
+    path: '/flyout-fixed-footer',
+    element: <FlyoutFixedFooterPage />,
+  },
+  {
+    name: 'Flyout Fixed Footer Header',
+    path: '/flyout-fixed-footer-header',
+    element: <FlyoutFixedFooterHeaderPage />,
+  },
+  {
+    name: 'Flyout Fixed Footer Header Scrollable Content',
+    path: '/flyout-fixed-footer-header-scrollable-content',
+    element: <FlyoutFixedFooterHeaderScrollableContentPage />,
+  },
+  {
+    name: 'Flyout Fixed Footer Header Subfooter',
+    path: '/flyout-fixed-footer-header-subfooter',
+    element: <FlyoutFixedFooterHeaderSubfooterPage />,
+  },
+  {
+    name: 'Flyout Fixed Footer Header Subfooter Scrollable Content',
+    path: '/flyout-fixed-footer-header-subfooter-scrollable-content',
+    element: <FlyoutFixedFooterHeaderSubfooterScrollableContentPage />,
+  },
+  {
+    name: 'Flyout Fixed Footer Scrollable Content',
+    path: '/flyout-fixed-footer-scrollable-content',
+    element: <FlyoutFixedFooterScrollableContentPage />,
+  },
+  {
+    name: 'Flyout Fixed Footer Scrollable Content Scrolled',
+    path: '/flyout-fixed-footer-scrollable-content-scrolled',
+    element: <FlyoutFixedFooterScrollableContentScrolledPage />,
+  },
+  {
+    name: 'Flyout Fixed Footer Subfooter',
+    path: '/flyout-fixed-footer-subfooter',
+    element: <FlyoutFixedFooterSubfooterPage />,
+  },
+  {
+    name: 'Flyout Fixed Footer Subfooter Scrollable Content',
+    path: '/flyout-fixed-footer-subfooter-scrollable-content',
+    element: <FlyoutFixedFooterSubfooterScrollableContentPage />,
+  },
+  {
+    name: 'Flyout Fixed Footer Subfooter Scrollable Content Scrolled',
+    path: '/flyout-fixed-footer-subfooter-scrollable-content-scrolled',
+    element: <FlyoutFixedFooterSubfooterScrollableContentScrolledPage />,
+  },
+  {
     name: 'Flyout Footer',
     path: '/flyout-footer',
     element: <FlyoutFooterPage />,
@@ -358,6 +426,11 @@ export const generatedRoutes: RouteType[] = [
     name: 'Flyout Footer Scrollable Content',
     path: '/flyout-footer-scrollable-content',
     element: <FlyoutFooterScrollableContentPage />,
+  },
+  {
+    name: 'Flyout Footer Scrolled',
+    path: '/flyout-footer-scrolled',
+    element: <FlyoutFooterScrolledPage />,
   },
   {
     name: 'Flyout Grid',
@@ -553,6 +626,11 @@ export const generatedRoutes: RouteType[] = [
     name: 'Modal Disable Close Button',
     path: '/modal-disable-close-button',
     element: <ModalDisableCloseButtonPage />,
+  },
+  {
+    name: 'Modal Footer Scrolled',
+    path: '/modal-footer-scrolled',
+    element: <ModalFooterScrolledPage />,
   },
   {
     name: 'Modal Fullscreen',