+++ packages/components-react/src/routes.tsx
@@ -137,6 +137,11 @@ export const routes: RouteType[] = [
     element: <fromExamples.FlyoutExamplePage />,
   },
   {
+    name: 'Flyout Example Conditional',
+    path: '/flyout-example-conditional',
+    element: <fromExamples.FlyoutExampleConditionalPage />,
+  },
+  {
     name: 'Flyout Multilevel Example',
     path: '/flyout-multilevel-example',
     element: <fromExamples.FlyoutMultilevelExamplePage />,
@@ -152,6 +157,11 @@ export const routes: RouteType[] = [
     element: <fromExamples.FlyoutMultilevelExampleCustomContentPage />,
   },
   {
+    name: 'Hidden Attribute Example',
+    path: '/hidden-attribute-example',
+    element: <fromExamples.HiddenAttributeExamplePage />,
+  },
+  {
     name: 'Inline Notification Example Action Button',
     path: '/inline-notification-example-action-button',
     element: <fromExamples.InlineNotificationExampleActionButtonPage />,
@@ -165,6 +175,11 @@ export const routes: RouteType[] = [
     name: 'Modal Example Accessibility',
     path: '/modal-example-accessibility',
     element: <fromExamples.ModalExampleAccessibilityPage />,
+  },
+  {
+    name: 'Modal Example Conditional',
+    path: '/modal-example-conditional',
+    element: <fromExamples.ModalExampleConditionalPage />,
   },
   {
     name: 'Multi-Select Example',