+++ packages/components-vue/src/examples/CheckboxExample.vue
@@ -13,6 +13,7 @@
   <form @submit.prevent="onSubmit">
     <PCheckbox :name="'some-name'" :label="'Some Label'" :value="'some-value'" />
     <button type="submit">Submit</button>
+    <button type="reset">Reset</button>
   </form>
 
   <PText>Last submitted data: {{ lastSubmittedData }}</PText>