+++ packages/components/scripts/wrapper-generator/UXPinReactWrapperGenerator.ts
@@ -13,18 +13,35 @@ const addNestedIndentation = (x: string): string => `  ${x}`;
 
 export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
   protected projectDir = 'uxpin-wrapper';
+  protected hiddenComponents: TagName[] = [];
 
   constructor() {
     super();
     this.ignoreComponents = [
       ...this.ignoreComponents,
+      'p-canvas',
+      'p-checkbox-wrapper',
       'p-content-wrapper',
+      'p-fieldset-wrapper',
       'p-flex',
       'p-flex-item',
+      'p-flyout-multilevel',
+      'p-flyout-multilevel-item',
       'p-grid',
       'p-grid-item',
+      'p-headline',
+      'p-link-social',
+      'p-marque',
       'p-pagination',
+      'p-select-wrapper',
+      'p-textarea-wrapper',
     ];
+
+    // components which should be generated and hidden in uxpin editor
+    this.hiddenComponents = [
+        'p-text-field-wrapper',
+        'p-radio-button-wrapper',
+    ]
   }
 
   public getComponentFileName(component: TagName): string {
@@ -96,12 +113,17 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
     }
 
     // add onClick prop for marque, buttons and links, but not button-group
-    else if (!!component.match(/(button|link|marque|stepper-horizontal-item|tag-dismissible)(?!-group)/)) {
+    else if (!!component.match(/(button|link|marque|stepper-horizontal-item|tag-dismissible|crest)(?!-group)/)) {
       props = addProp(props, 'onClick?: (e: MouseEvent) => void;');
     }
 
     // remove BreakpointCustomizable types since designers can't use JSON
     props = props.replace(/BreakpointCustomizable<(.*)>/g, '$1');
+
+
+    // hidden uxpin props which allows updating property from library level in uxpin editor
+    props = addProp(props, '/** @uxpinignoreprop */ \n  uxpinOnChange: (prevValue: any, nextValue: any, propertyName: string) => void;');
+
 
     // remove useless props
     if (component === 'p-marque') {
@@ -120,8 +142,20 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
     // add uxpinignoreprop annotations
     if (component === 'p-modal') {
       props = addUxPinIgnorePropAnnotation(props, 'open');
-    } else if (component === 'p-link' || component === 'p-link-pure' || component === 'p-link-social') {
+    } else if (component === 'p-link' || component === 'p-link-pure' || component === 'p-link-social'  || component === 'p-crest') {
       props = addUxPinIgnorePropAnnotation(props, 'href');
+      props = addUxPinIgnorePropAnnotation(props, 'target');
+    } else if (component === 'p-banner') {
+      props = addUxPinIgnorePropAnnotation(props, 'width');
+    } else if (component === 'p-button' || component === 'p-button-pure') {
+      props = addUxPinIgnorePropAnnotation(props, 'name');
+      props = addUxPinIgnorePropAnnotation(props, 'value');
+    } else if (component === 'p-icon') {
+      props = addUxPinIgnorePropAnnotation(props, 'lazy');
+    } else if (component === 'p-model-signature') {
+      props = addUxPinIgnorePropAnnotation(props, 'fetchPriority');
+      props = addUxPinIgnorePropAnnotation(props, 'lazy');
+
     }
 
     // add uxpinbind annotations
@@ -137,6 +171,8 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
     } else if (component === 'p-tabs-bar') {
       props = addUxPinBindAnnotation(props, 'activeTabIndex', 'onUpdate', 'activeTabIndex');
       props = addUxPinBindAnnotation(props, 'activeTabIndex', 'onTabChange', 'activeTabIndex');
+    } else if (component === 'p-select' || component === 'p-multi-select') {
+      props = addUxPinBindAnnotation(props, 'value', 'onUpdate', 'value');
     }
 
     return props;
@@ -221,6 +257,67 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
       cleanedComponent = removeDestructuredProp(cleanedComponent, 'target');
     }
 
+    if (component === 'p-pin-code') {
+      cleanedComponent = cleanedComponent.replace(
+          'useEventCallback(elementRef, \'update\', onUpdate as any);',
+          [
+              'const eventCallback = (e:Event) => {',
+              '       rest.uxpinOnChange(value, (e as CustomEvent<PinCodeUpdateEventDetail>).detail.value, \'value\');',
+              '       if (onUpdate) {',
+              '         onUpdate(e as CustomEvent<PinCodeUpdateEventDetail>);',
+              '       }',
+              '    }',
+              '    useEventCallback(elementRef, \'update\', eventCallback);',
+          ].join('\n')
+      )
+    }
+
+    if (['p-flyout', 'p-modal', 'p-banner'].includes(component)) {
+      cleanedComponent = cleanedComponent.replace(
+          'useEventCallback(elementRef, \'dismiss\', onDismiss as any);',
+          [
+            'const dismissCallback = (e:Event) => {',
+            '       rest.uxpinOnChange(open, false, \'open\');',
+            '       if (onDismiss) {',
+            '         onDismiss(e as CustomEvent<void>);',
+            '       }',
+            '    }',
+            '    useEventCallback(elementRef, \'dismiss\', dismissCallback);',
+          ].join('\n')
+      )
+    }
+
+    // make crest and link-pure anchor if onClick is defined
+    if (component === 'p-crest' || component === 'p-link-pure') {
+      cleanedComponent = cleanedComponent.replace(
+          'const props = {',
+          [
+            '',
+            'useBrowserLayoutEffect(() => {',
+            '  const { current } = elementRef;',
+            '  (current as any).href = rest.onClick ? \'#\' : undefined;',
+            '}, [rest.onClick]);',
+            '',
+            'const props = {',
+          ].join('\n    ')
+      )
+    }
+
+    if (component === 'p-inline-notification') {
+      cleanedComponent = cleanedComponent.replace(
+          'useEventCallback(elementRef, \'dismiss\', onDismiss as any);',
+          [
+            'const dismissCallback = (e:Event) => {',
+            '       rest.uxpinOnChange(\`visible\`, \'hidden\', \'stateIa\');',
+            '       if (onDismiss) {',
+            '         onDismiss(e as CustomEvent<void>);',
+            '       }',
+            '    }',
+            '    useEventCallback(elementRef, \'dismiss\', dismissCallback);',
+          ].join('\n')
+      )
+    }
+
     // cast BreakpointCustomizable default prop values to any because BreakpointCustomizable types are removed for uxpin
     extendedProps
       .filter((prop) => prop.isDefaultValueComplex && prop.defaultValue.match(/\bbase\b/))
@@ -234,15 +331,19 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
   public shouldGenerateFolderPerComponent(component: TagName): boolean {
     switch (component) {
       case 'p-accordion':
+      case 'p-banner':
       case 'p-button-group':
       case 'p-button-tile':
-      case 'p-checkbox-wrapper':
+      case 'p-carousel':
+      case 'p-checkbox':
       case 'p-fieldset':
       case 'p-link-tile':
       case 'p-link-tile-model-signature':
+      case 'p-multi-select':
       case 'p-modal':
       case 'p-radio-button-wrapper':
       case 'p-segmented-control':
+      case 'p-select':
       case 'p-select-wrapper':
       case 'p-stepper-horizontal':
       case 'p-text-field-wrapper':
@@ -271,6 +372,9 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
         props: { heading: 'Heading' },
         children: '<Text uxpId="accordion-text" children="Content" />',
       },
+      'p-banner': {
+        props: { heading: 'Heading', description: 'Description', open: true, },
+      },
       'p-button-group': {
         children: [
           '<Button variant="primary" uxpId="button-primary" />',
@@ -281,13 +385,19 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
         props: { label: 'Some label', description: 'Some description' },
         children: '<DummyImg uxpId="dummy-img" />',
       },
-      'p-checkbox-wrapper': {
-        props: { label: 'CheckboxWrapper' },
-        children: '<DummyCheckbox uxpId="dummy-checkbox" />',
-        formComponent: {
-          name: 'Checkbox',
-          extraProps: { label: 'My Checkbox', checked: true },
-        },
+      'p-carousel': {
+        props: { heading: 'Some heading' },
+        children: [
+          '<DummyDiv uxpId="dummy-div-1" uxpinCustomStyles={{ display: \'flex\', alignItems: \'center\', justifyContent: \'center\', background: \'#00b0f4\', height: 150 }} children="Slide 1" />',
+          '<DummyDiv uxpId="dummy-div-2" uxpinCustomStyles={{ display: \'flex\', alignItems: \'center\', justifyContent: \'center\', background: \'#00b0f4\', height: 150 }}  children="Slide 2" />',
+          '<DummyDiv uxpId="dummy-div-3" uxpinCustomStyles={{ display: \'flex\', alignItems: \'center\', justifyContent: \'center\', background: \'#00b0f4\', height: 150 }}  children="Slide 3" />',
+          '<DummyDiv uxpId="dummy-div-4" uxpinCustomStyles={{ display: \'flex\', alignItems: \'center\', justifyContent: \'center\', background: \'#00b0f4\', height: 150 }}  children="Slide 4" />',
+          '<DummyDiv uxpId="dummy-div-5" uxpinCustomStyles={{ display: \'flex\', alignItems: \'center\', justifyContent: \'center\', background: \'#00b0f4\', height: 150 }}  children="Slide 5" />',
+
+        ].join(glue),
+      },
+      'p-checkbox': {
+        props: { label: 'label' },
       },
       'p-fieldset': {
         props: { label: 'Fieldset' },
@@ -304,11 +414,29 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
           '<Link slot="secondary" variant="secondary" theme="dark" href="#" uxpId="link-secondary">Some link</Link>', // we need to set variant and theme props for uxpin editor to display the right config
         ].join(glue),
       },
+      'p-multi-select': {
+        props: { name:'options', label:'Some Label' },
+        children: [
+          '<Optgroup uxpId="group-1" label="Some optgroup label 1">',
+          ' <MultiSelectOption uxpId="opt-1" value="a">Option A</MultiSelectOption>',
+          ' <MultiSelectOption uxpId="opt-2" value="b">Option B</MultiSelectOption>',
+          ' <MultiSelectOption uxpId="opt-3" value="c">Option C</MultiSelectOption>',
+          ' <MultiSelectOption uxpId="opt-4" value="d">Option D</MultiSelectOption>',
+          ' <MultiSelectOption uxpId="opt-5" value="e">Option E</MultiSelectOption>',
+          ' <MultiSelectOption uxpId="opt-6" value="f">Option F</MultiSelectOption>',
+          '</Optgroup>',
+          '<Optgroup uxpId="group-3" label="Some optgroup label 2">',
+          ' <MultiSelectOption uxpId="opt-7"  value="g">Option G</MultiSelectOption>',
+          ' <MultiSelectOption uxpId="opt-8"  value="h">Option H</MultiSelectOption>',
+          ' <MultiSelectOption uxpId="opt-9"  value="i">Option I</MultiSelectOption>',
+          '</Optgroup>',
+        ].join(glue),
+      },
       'p-modal': {
         props: { heading: 'Heading', open: true },
         children: [
           '<Text uxpId="modal-text">Some Content</Text>',
-          '<ButtonGroup uxpId="modal-button-group" >',
+          '<ButtonGroup slot="footer" uxpId="modal-button-group" >',
           ...[
             '<Button uxpId="modal-button-1" children="Save" />',
             '<Button uxpId="modal-button-2" variant="tertiary" children="Close" />',
@@ -335,14 +463,18 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
           )
           .join(glue),
       },
+      'p-select': {
+        props: { name:'options', label:'Some Label', description: 'Some description', value:'a'  },
+        children: [
+            '<SelectOption uxpId="opt-1" value="a">Option A</SelectOption>',
+            '<SelectOption uxpId="opt-2" value="b">Option B</SelectOption>',
+            '<SelectOption uxpId="opt-3" value="c">Option C</SelectOption>',
+        ].join(glue),
+      },
       'p-select-wrapper': {
         props: { label: 'SelectWrapper' },
         children:
           '<DummySelect uxpId="dummy-select" options={Array.from(Array(3)).map((_, i) => `Option ${i + 1}`)} />',
-        formComponent: {
-          name: 'SelectWrapperDummy',
-          extraProps: { label: 'My Select', options: ['Option 1', 'Option 2', 'Option 3'] },
-        },
       },
       'p-stepper-horizontal': {
         children: [
@@ -362,10 +494,6 @@ export class UXPinReactWrapperGenerator extends ReactWrapperGenerator {
       'p-textarea-wrapper': {
         props: { label: 'TextareaWrapper' },
         children: '<DummyTextarea uxpId="dummy-textarea" />',
-        formComponent: {
-          name: 'Textarea',
-          extraProps: { label: 'My Textarea' },
-        },
       },
       'p-table': {
         children: [
@@ -531,7 +659,12 @@ export default <${formComponentName} ${stringifiedProps} />;
 
   private generateUXPinConfigFile(): AdditionalFile {
     const componentsBasePath = 'src/lib/components/';
-    const componentPaths = this.relevantComponentTagNames
+    const uxpinComponents = [
+      `'src/form/RadioButton/RadioButton.tsx'`,
+      `'src/form/TextField/TextField.tsx'`,
+    ];
+    const componentPaths = [...this.relevantComponentTagNames
+      .filter((component) => !this.hiddenComponents.includes(component))
       .map((component) => {
         const componentSubDir = this.shouldGenerateFolderPerComponent(component)
           ? this.stripFileExtension(component) + '/'
@@ -539,8 +672,13 @@ export default <${formComponentName} ${stringifiedProps} />;
         const fileName = this.getComponentFileName(component);
         return `${componentsBasePath}${componentSubDir}${fileName}`;
       })
-      .map((path) => `'${path}'`)
-      .join(',\n          ');
+      .map((path) => `'${path}'`),
+      ...uxpinComponents
+    ].sort((componentA, componentB) => (
+      componentA.split('/').pop().toLowerCase()
+          .localeCompare(componentB.split('/').pop().toLowerCase())
+    )).join(',\n          ');
+
 
     const content = `module.exports = {
   components: {
@@ -553,19 +691,32 @@ export default <${formComponentName} ${stringifiedProps} />;
         ],
       },
       {
-        name: 'Form components',
-        include: ['src/form/*/*.tsx'],
-      },
-      {
         name: 'Dummy',
-        include: ['src/dummy/*.tsx'],
+        include: [
+         'src/dummy/DummyButton.tsx',
+         'src/dummy/DummyImg.tsx',
+         'src/dummy/DummyLink.tsx',
+         'src/dummy/DummySpan.tsx',
+         'src/dummy/DummyDiv.tsx'
+        ],
       },
     ],
     wrapper: 'src/UXPinWrapper.tsx',
     webpackConfig: 'webpack.config.js',
   },
   name: 'Porsche Design System',
-  settings: { useUXPinProps: true, useFitToContentAsDefault: true },
+  settings: { 
+    useUXPinProps: true, 
+    useFitToContentAsDefault: true,
+    propertyConfigurations: {
+      Flyout: {
+        open: { disabled: true, context: 'canvas', value: false, },
+      },
+      Modal: {
+        open: { disabled: true, context: 'canvas', value: false },
+      }
+    }
+  },
 };`;
 
     return { name: 'uxpin.config.js', relativePath: '../../..', content };