+++ packages/components/src/components/textarea/textarea.tsx
@@ -140,6 +140,7 @@ export class Textarea {
 
   @AttachInternals() private internals: ElementInternals;
 
+  private defaultValue: string;
   private textAreaElement: HTMLTextAreaElement;
   private counterElement: HTMLSpanElement;
   private hasCounter: boolean;
@@ -165,14 +166,21 @@ export class Textarea {
   }
 
   public componentWillLoad(): void {
+    this.defaultValue = this.value;
     this.updateCounterVisibility();
   }
 
   public formResetCallback(): void {
-    this.internals.setValidity({});
-    this.internals.setFormValue('');
-    this.textAreaElement.value = '';
-    this.value = '';
+    this.internals.setFormValue(this.defaultValue);
+    this.value = this.defaultValue;
+  }
+
+  public formDisabledCallback(disabled: boolean): void {
+    this.disabled = disabled;
+  }
+
+  public formStateRestoreCallback(state: string): void {
+    this.value = state;
   }
 
   public componentShouldUpdate(newVal: unknown, oldVal: unknown): boolean {
@@ -186,6 +194,11 @@ export class Textarea {
     if (this.hasCounter) {
       this.setCounterAriaTextDebounced();
     }
+    this.internals.setValidity(
+      this.textAreaElement.validity,
+      this.textAreaElement.validationMessage,
+      this.textAreaElement
+    );
   }
 
   public render(): JSX.Element {